+++ controllers/index.js
@@ -4,13 +4,13 @@ const homeRoutes = require('./home-routes.js');
 // const apiRoutes = require('./apii');
 const apiRoutes = require('./api');
 // const homeRoutes = require('./homeRoutes');
-const authRoutes = require('./auths');
-// const withAuth = require('../utils/auth');
+const authRoutes = require('./api/auths');
+const withAuth = require('../utils/auth');
 
 
 // might have to refactor this.. currently set to /api routes require login.
 router.use('/', homeRoutes);
-router.use('/auth', authRoutes);
+// router.use('/auth', authRoutes);
 router.use('/api', apiRoutes);
 
 router.use((req, res) => {
